jni: refine JNI and try to improve stability of llava inference and stablediffusion inference
